Student - fixed Passed_Credit_Iterator
The problem was in using `exam_list[i].BinarySearch` to look for an element with a given Exam_Name. BinarySearch for typified list requires to compare Exams to Exams and can not use my custom comparator. Therefor I had to switch to exam_list.FindIndex which uses a custom predicate to search for an element.